test: cleaned up probes for g++

+++ test/g++.dg-torture/dmce-prolog-test-D5.c
@@ -1,7 +1,6 @@
 /* Test prolog for code trace and data trace with 5 parameters */
 #ifndef __DMCE_PROBE_FUNCTION__HEADER__
 #define __DMCE_PROBE_FUNCTION__HEADER__
-#include <stdint.h>
-static void dmce_probe_body(unsigned int dmce_probenbr, uint64_t dmce_param_a, uint64_t dmce_param_b, uint64_t dmce_param_c, uint64_t dmce_param_d, uint64_t dmce_param_e);
+static void dmce_probe_body(unsigned long dmce_probenbr, unsigned long dmce_param_a, unsigned long dmce_param_b, unsigned long dmce_param_c, unsigned long dmce_param_d, unsigned long dmce_param_e);
 #define DMCE_PROBE(dmce_probenbr, dmce_param_a, dmce_param_b, dmce_param_c, dmce_param_d, dmce_param_e) (dmce_probe_body(dmce_probenbr, dmce_param_a, dmce_param_b, dmce_param_c, dmce_param_d, dmce_param_e))
 #endif
